Fix stage logs
