+++ app/styles/stage-details.less
@@ -20,3 +20,13 @@
   -webkit-transition: opacity .1s ease;
   transition: opacity .1s ease;
 }
+
+.stage-details > .segment  > .menu {
+  border-radius: 0;
+  border-width: 1px 0 1px 0;
+  box-shadow: none;
+}
+
+.stage-details > .segment > .segment.tab {
+  border-width: 0;
+}
