+++ app/templates/components/build-stage.hbs
@@ -1,0 +1,2 @@
+<i class="{{statusIcon}} icon"></i>
+<div class="label">{{model.name}}</div>
